refactor: move exception classes to the flexmeasures.exceptions (#80)
