+++ src/flexmeasures_client/client.py
@@ -19,6 +19,13 @@ from flexmeasures_client.constants import (
     CONTENT_TYPE_HEADERS,
     ENTITY_ADDRESS_PLACEHOLDER,
 )
+from flexmeasures_client.exceptions import (
+    ContentTypeError,
+    EmailValidationError,
+    EmptyPasswordError,
+    WrongAPIVersionError,
+    WrongHostError,
+)
 from flexmeasures_client.response_handling import (
     check_content_type,
     check_for_status,
@@ -690,43 +697,3 @@ class FlexMeasuresClient:
                 f"Power conversion from {from_unit} to {to_unit} is not supported."
             )
         return values
-
-
-class ContentTypeError(Exception):
-    """Raised when the response from the API is not in the expected format"""
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
-
-
-class WrongHostError(Exception):
-    """Raised when the host includes the scheme (http:// or https://)"""
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
-
-
-class EmptyPasswordError(Exception):
-    """Raised when the password is empty"""
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
-
-
-class EmailValidationError(Exception):
-    """Raised when the email is not in the correct format"""
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
-
-
-class WrongAPIVersionError(Exception):
-    """Raised when the API version is not in the versions list"""
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
