--- src/flexmeasures_client/s2/exceptions.py
@@ -1,2 +1,0 @@
-class WrongCredentialsException(Exception):
-    pass
